requirements.txt
Flask==3.0.0
gunicorn==21.2.0
lxml==6.0.2
pillow==12.0.0
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
